DEPR: Renamed pybaselines.window to pybaselines.smooth
Renamed the module to make its usage more clear. For now, pybaselines.window will redirect to pybaselines.smooth. In a later version, it will be deprecated, and later removed. Changed all docs and examples.
