+++ docs/_templates/autoapi/python/module.rst
@@ -5,6 +5,11 @@
 {{ obj.name }}
 {{ "=" * obj.name|length }}
 
+{% if obj.name == "pybaselines.window" %}
+{{ "This module was renamed in version 0.6.0. Use :mod:`pybaselines.smooth` instead." }}
+
+{% endif %}
+
 .. py:module:: {{ obj.name }}
